% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma.R
\name{fdr_level}
\alias{fdr_level}
\title{Get the FDR threshold from a set of p-values}
\usage{
fdr_level(pval, al = 0.05)
}
\arguments{
\item{pval}{}
}
\description{
Get the FDR threshold from a set of p-values
}
\details{
The function will do a p-value adjustment using p.adjust(method='fdr') and
then will search for the p-value in the vector provided that corresponds to
that element that is below the alpha level. If there is none, it will report
the minimum p-value after the FDR transformation, which will be above the
alpha level and then indicates that none of the provided p-values would be
significant after the FDR correction.
}
