% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma.R
\name{genplink}
\alias{genplink}
\title{Create a directory with all files needed for GEMMA
Generates the plink necessary fam file to run gemma}
\usage{
genplink(data, id = "id", phenotype, naval = "-9", out = "test",
  path = "../gemma/", famfile = "515g.fam", bimfile = "515g.bim",
  bedfile = "515g.bed", cleardir = F, makerelative = F)
}
\description{
Create a directory with all files needed for GEMMA
Generates the plink necessary fam file to run gemma
}
