% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma.R
\name{run_gemma}
\alias{run_gemma}
\title{Call GEMMA from within R}
\usage{
run_gemma(plinkbase, plinkfolder = ".", out = ".", type = "bslmm",
  maf = 0, background = TRUE, dryrun = FALSE)
}
\arguments{
\item{dryrun}{}
}
\description{
Call GEMMA from within R
}
