% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_MCMC_GWA}
\alias{run_MCMC_GWA}
\title{run}
\usage{
run_MCMC_GWA(sdat, Go, epi = 1, simumodes, fitnessmodes, priormodes,
  proposalmodes, gwatypes, iterations = 500, burnin = 0.1, donap = TRUE,
  dogwa = TRUE, doplot = TRUE, pdfname = NULL)
}
\arguments{
\item{doplot}{}
}
\description{
run
}
