% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate}
\usage{
simulate_data(Go, p = 0.1, b = 0.1, a = 0, mu = 1, epi = 1,
  svar = 0.5, ss = 0.8, m = 10, trainingp = 1, prophidden = 0)
}
\arguments{
\item{epi}{}
}
\description{
Simulate
}
