% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomesio.R
\name{XBMsimulate}
\alias{XBMsimulate}
\title{Simulate a big genome matrix with filebacked system}
\usage{
XBMsimulate(backingpath = "databig", filename = "example", n, m,
  force = F)
}
\arguments{
\item{force}{}
}
\description{
Simulate a big genome matrix with filebacked system
}
